+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Fix the JDBC driver version returned by the methods of the classes `CassandraDriver` and `CassandraDatabaseMetaData` 
+  to be consistent with the version of the JDBC wrapper artifact (see issue 
+  [#19](https://github.com/ing-bank/cassandra-jdbc-wrapper/issues/19)).
+
 ## 4.8.0 - 2023-01-12
 ### Added
 - Implement the methods `getMetaData()` and `getParameterMetaData()` into the implementation class